#18 HammerTime Marathon Bot:
Moved HammerTime bot to separate module, added common module and improved the structure of the project
Migrated Sheets Client to Kotlin
