+++ commons/src/main/kotlin/press/lis/greeb/spreadsheets/SheetsClient.kt
@@ -1,0 +1,66 @@
+/**
+ * @author Aleksandr Eliseev
+ */
+package press.lis.greeb.spreadsheets
+
+import com.google.api.client.auth.oauth2.Credential
+import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp
+import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver
+import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow
+import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets
+import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
+import com.google.api.client.http.javanet.NetHttpTransport
+import com.google.api.client.json.JsonFactory
+import com.google.api.client.json.jackson2.JacksonFactory
+import com.google.api.client.util.store.FileDataStoreFactory
+import com.google.api.services.sheets.v4.Sheets
+import com.google.api.services.sheets.v4.SheetsScopes
+import java.io.File
+import java.io.FileNotFoundException
+import java.io.IOException
+import java.io.InputStreamReader
+
+object SheetsClient {
+    private const val APPLICATION_NAME = "Google Sheets API Java Quickstart"
+    private val JSON_FACTORY: JsonFactory = JacksonFactory.getDefaultInstance()
+    private const val TOKENS_DIRECTORY_PATH = "tokens"
+
+    /**
+     * Global instance of the scopes required by this quickstart.
+     * If modifying these scopes, delete your previously saved tokens/ folder.
+     */
+    private val SCOPES = listOf(SheetsScopes.SPREADSHEETS)
+    private const val CREDENTIALS_FILE_PATH = "/credentials.json"
+
+    /**
+     * Creates an authorized Credential object.
+     *
+     * @param HTTP_TRANSPORT The network HTTP Transport.
+     * @return An authorized Credential object.
+     * @throws IOException If the credentials.json file cannot be found.
+     */
+    @Throws(IOException::class)
+    private fun getCredentials(HTTP_TRANSPORT: NetHttpTransport): Credential {
+        // Load client secrets.
+        val `in` = SheetsClient::class.java.getResourceAsStream(CREDENTIALS_FILE_PATH)
+                ?: throw FileNotFoundException("Resource not found: $CREDENTIALS_FILE_PATH")
+        val clientSecrets = GoogleClientSecrets.load(JSON_FACTORY, InputStreamReader(`in`))
+
+        // Build flow and trigger user authorization request.
+        val flow = GoogleAuthorizationCodeFlow.Builder(
+                        HTTP_TRANSPORT, JSON_FACTORY, clientSecrets, SCOPES)
+                .setDataStoreFactory(FileDataStoreFactory(File(TOKENS_DIRECTORY_PATH)))
+                .setAccessType("offline")
+                .build()
+        val receiver = LocalServerReceiver.Builder().setPort(8888).build()
+        return AuthorizationCodeInstalledApp(flow, receiver).authorize("user")
+    }
+
+    val sheetService: Sheets
+        get() {
+            val httpTransport = GoogleNetHttpTransport.newTrustedTransport()
+            return Sheets.Builder(httpTransport, JSON_FACTORY, getCredentials(httpTransport))
+                    .setApplicationName(APPLICATION_NAME)
+                    .build()
+        }
+}
