+++ hammertime-bot/src/main/kotlin/press/lis/greeb/hammertime/HammerTimeMarathonConstants.kt
@@ -1,0 +1,41 @@
+package press.lis.greeb.hammertime
+
+/**
+ * @author Aleksandr Eliseev
+ */
+object HammerTimeMarathonConstants {
+    val links = mapOf(
+            "0" to "https://lesswrong.ru/w/%D0%9C%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%B8_%D0%B8_%D0%B3%D0%B2%D0%BE%D0%B7%D0%B4%D0%B8",
+            "1" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_1_%D0%9E%D1%85%D0%BE%D1%82%D0%B0_%D0%B7%D0%B0_%D0%B1%D0%B0%D0%B3%D0%B0%D0%BC%D0%B8",
+            "2" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_2_%D0%99%D0%BE%D0%B4%D0%B0-%D1%82%D0%B0%D0%B9%D0%BC%D0%B5%D1%80",
+            "3" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_3_%D0%9F%D0%A2%D0%94",
+            "4" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_4_%D0%94%D0%B8%D0%B7%D0%B0%D0%B9%D0%BD",
+            "5" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_5_%D0%A0%D0%B0%D1%81%D1%88%D0%B8%D1%80%D0%B5%D0%BD%D0%B8%D0%B5_%D0%B7%D0%BE%D0%BD%D1%8B_%D0%BA%D0%BE%D0%BC%D1%84%D0%BE%D1%80%D1%82%D0%B0",
+            "6" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_6_%D0%9C%D0%B0%D0%BD%D1%82%D1%80%D1%8B",
+            "7" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_7_%D0%A4%D0%B0%D0%BA%D1%82%D0%BE%D1%80%D0%B8%D0%B7%D0%B0%D1%86%D0%B8%D1%8F_%D0%B8%D0%B7%D0%B1%D0%B5%D0%B3%D0%B0%D0%BD%D0%B8%D0%B9",
+            "8" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_8_%D0%92%D0%B5%D1%80%D0%B0_%D0%B2_%D0%BD%D0%B5%D0%B2%D0%BE%D0%B7%D0%B2%D1%80%D0%B0%D1%82%D0%BD%D1%8B%D0%B5_%D0%B8%D0%B7%D0%B4%D0%B5%D1%80%D0%B6%D0%BA%D0%B8",
+            "9" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_9_%D0%9A%D0%B0%D0%BB%D0%B8%D0%B1%D1%80%D0%BE%D0%B2%D0%BA%D0%B0_%D0%B2%D1%80%D0%B5%D0%BC%D0%B5%D0%BD%D0%B8",
+            "10" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_10_%D0%9C%D1%91%D1%80%D1%84%D0%B8%D0%B4%D0%B6%D0%B8%D1%82%D1%81%D1%83",
+            "11" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_11_%D0%9E%D1%85%D0%BE%D1%82%D0%B0_%D0%BD%D0%B0_%D0%B1%D0%B0%D0%B3%D0%B8_2",
+            "12" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_12_%D0%99%D0%BE%D0%B4%D0%B0-%D1%82%D0%B0%D0%B9%D0%BC%D0%B5%D1%80%D1%8B_2",
+            "13" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_13_%D0%9F%D0%A2%D0%94_2",
+            "14" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_14_%D0%94%D0%B8%D0%B7%D0%B0%D0%B9%D0%BD_2",
+            "15" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_15_%D0%A0%D0%B0%D0%97%D0%BE%D0%9A_2",
+            "16" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_16_%D0%A2%D1%80%D0%B8_%D0%BC%D0%B8%D0%BD%D0%B8%D0%B0%D1%82%D1%8E%D1%80%D1%8B",
+            "17" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_17_%D0%A4%D0%BE%D0%BA%D1%83%D1%81%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%B8%D0%B5",
+            "18" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_18_%D0%A4%D0%B0%D0%BA%D1%82%D0%BE%D1%80%D0%B8%D0%B7%D0%B0%D1%86%D0%B8%D1%8F_%D1%86%D0%B5%D0%BB%D0%B5%D0%B9",
+            "19" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_19_TDT_%D0%B4%D0%BB%D1%8F_%D0%BB%D1%8E%D0%B4%D0%B5%D0%B9",
+            "20" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_20_%D0%94%D1%80%D1%83%D0%B6%D0%B1%D0%B0",
+            "21" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_21_%D0%9E%D1%85%D0%BE%D1%82%D0%B0_%D0%B7%D0%B0_%D0%B1%D0%B0%D0%B3%D0%B0%D0%BC%D0%B8_3",
+            "22" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_22_%D0%99%D0%BE%D0%B4%D0%B0-%D1%82%D0%B0%D0%B9%D0%BC%D0%B5%D1%80%D1%8B_3",
+            "23" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_23_%D0%9F%D0%A2%D0%94_3",
+            "24" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_24_%D0%94%D0%B8%D0%B7%D0%B0%D0%B9%D0%BD_3",
+            "25" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_25_%D0%A0%D0%B0%D0%97%D0%BE%D0%9A_3",
+            "26" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_26_%D0%A2%D0%B8%D1%88%D0%B8%D0%BD%D0%B0",
+            "27" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_27_%D0%92%D0%BD%D1%83%D1%82%D1%80%D0%B5%D0%BD%D0%BD%D0%B8%D0%B9_%D0%BA%D0%BE%D1%80%D0%B5%D0%BD%D1%8C_%D1%80%D0%B0%D0%B7%D0%BD%D0%BE%D0%B3%D0%BB%D0%B0%D1%81%D0%B8%D1%8F",
+            "28" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_28_%D0%9F%D0%BE%D0%B2%D1%82%D0%BE%D1%80%D0%B5%D0%BD%D0%B8%D0%B5_%D1%80%D0%B5%D0%B4%D1%83%D0%BA%D1%86%D0%B8%D0%BE%D0%BD%D0%B8%D0%B7%D0%BC%D0%B0",
+            "29" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_29_%D0%A1%D1%82%D1%80%D0%B0%D1%82%D0%B5%D0%B3%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_%D1%83%D1%80%D0%BE%D0%B2%D0%B5%D0%BD%D1%8C",
+            "30" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%B4%D0%B5%D0%BD%D1%8C_30_%D0%A4%D0%B8%D0%BD%D0%B0%D0%BB%D1%8C%D0%BD%D1%8B%D0%B9_%D1%8D%D0%BA%D0%B7%D0%B0%D0%BC%D0%B5%D0%BD",
+            "31" to "https://lesswrong.ru/w/%D0%92%D1%80%D0%B5%D0%BC%D1%8F_%D0%BC%D0%BE%D0%BB%D0%BE%D1%82%D0%BA%D0%BE%D0%B2_%D0%BF%D0%BE%D1%81%D0%BB%D0%B5%D1%81%D0%BB%D0%BE%D0%B2%D0%B8%D0%B5")
+
+}
