--- src/main/resources/logback.xml
@@ -1,13 +1,0 @@
-<?xmlSettingsPath version="1.0" encoding="UTF-8"?>
-<configuration scan="true" scanPeriod="10 seconds">
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <!-- use IntelliJ Idea Grep Console to colorize logs -->
-            <pattern>%d{yyyy-MM-dd HH:mm:ss,SSS Z} [%highlight(%-5level)] [%cyan(%logger)] [%blue(%class{0})] %mdc: %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <root level="DEBUG">
-        <appender-ref ref="STDOUT" />
-    </root>
-</configuration>
